chore: Move on-premise pkg to utils
