+++ internal/utils/get_distro.go
@@ -1,4 +1,4 @@
-package onpremise
+package utils
 
 import (
 	"bufio"
