+++ internal/utils/install_dependencies.go
@@ -1,4 +1,4 @@
-package onpremise
+package utils
 
 import (
 	"runtime"
